M2Pro2
This assignment will require a good understanding of the skills learned in this module, but it will also require researching some concepts that might not have been covered in depth. You will be provided with some links to help with these topics, yet you might still need to research further. Well, because researching and learning new skills will always be a big part of your career, and if some are thinking of ChatGPT, then yes, you can use it. Especially since I would like for you to try getting this assignment done on your own, you are allowed to use it.